backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pgvector==0.3.6

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Validation & Settings ─────────────────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic[email]==2.10.3

# ── AI ────────────────────────────────────────────────────────────────────────
openai==1.57.4

# ── AWS / Storage ─────────────────────────────────────────────────────────────
boto3==1.35.88

# ── PDF Ingestion ─────────────────────────────────────────────────────────────
pypdf==5.1.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-multipart==0.0.20
python-dotenv==1.0.1
httpx==0.28.1
